refactor robot_functions.py
