+++ aisaac/scripts/robot_strategy.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env  python
+# coding:utf-8
+
+class RobotStrategy:
+    def __init__(self):
+        pass
+
+    def attack_startegy(self):
+        pass
