+++ aisaac/scripts/robot_utils.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env  python
+# coding:utf-8
+
+ROBOT_LOOP_RATE = 60.
