+++ aisaac/scripts/world_model.py
@@ -21,7 +21,8 @@ from aisaac.msg import Status
 
 WORLD_LOOP_RATE = 100.
 
-from world_model_functions import WorldState, Referee, Objects, DecisionMaker
+from world_model_functions import WorldState, Referee, DecisionMaker
+from objects import Objects
 import functions
 
 #import physics as phy
